+++ expenses/views/transactions.py
@@ -31,6 +31,7 @@ class TransactionGroupListView(ListView):
         # Filter expenses by the specified period
         queryset = (
             Transaction.objects.filter(period=self.period_id)
+            .exclude(account__name=settings.INVALID_ACCOUNT)
             .values("account")
             .annotate(total=Sum("local_amount"))
             .values("account__name", "account__id", "total")
@@ -48,6 +49,7 @@ class TransactionGroupListView(ListView):
         context["total"] = get_total_local_amount(Q(period=period))
         _last = (
             Transaction.objects.filter(period=self.period_id)
+            .exclude(account__name=settings.INVALID_ACCOUNT)
             .values("payment_date")
             .order_by("-payment_date")
             .first()
@@ -118,7 +120,9 @@ class TransactionDeleteView(DeleteView):
 
 class TransactionRemoveInvalidView(APIView):
     def get(self, request):
-        invalid_expenses = Transaction.objects.filter(account__name="Invalido")
+        invalid_expenses = Transaction.objects.filter(
+            account__name=settings.INVALID_ACCOUNT
+        )
         invalid_expenses.delete()
         return redirect("home")